The controller has been tested in works smoothly.
I have also done the write up on this controller, and this has been been done in the main file used for all the stm32 microcontroller file. This file has now exceeded 25MB , which is the maximum file size that can be uploaded on git, this I could not upload it, but I will save it in my local disk though.